requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-ckeditor==6.4.0
django-ckeditor-5==0.0.16
django-js-asset==2.0.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.0
pycodestyle==2.8.0
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
